requirements.txt
folium==0.14.0
numpy==1.24.2
pandas==1.5.3
plotly==5.13.1
streamlit==1.20.0
streamlit-folium==0.11.1
Pillow==9.0.1
matplotlib-inline==0.1.6